🔨 refactor: interface methods of measurement module now pass to common module
