+++ common/model/interface.go
