+++ common/usecase/interface.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/metalpoch/olt-blueprint/common/constants"
 	"github.com/metalpoch/olt-blueprint/common/entity"
+	"github.com/metalpoch/olt-blueprint/common/model"
 	"github.com/metalpoch/olt-blueprint/common/pkg/tracking"
-	"github.com/metalpoch/olt-blueprint/measurement/model"
-	"github.com/metalpoch/olt-blueprint/measurement/repository"
+	"github.com/metalpoch/olt-blueprint/common/repository"
 	"gorm.io/gorm"
 )
